+++ .website/tailwind.config.js
@@ -1,0 +1,9 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: ['components/**/*.vue', '**/*.md', '.vitepress/theme/*.vue'],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
+