+++ benchmarks/bin/benchmarks.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:benchmarks/benchmarks.dart' as benchmarks;
@@ -7,23 +8,25 @@ import 'package:benchmarks/shared/serinus_benchmark.dart';
 Map<String, Result?> results = {};
 
 Future<void> main(List<String> arguments) async {
-  results['shelf'] = await benchmarks.ShelfAppBenchmark().report();
   results['serinus'] = await benchmarks.SerinusAppBenchmark().report();
+  // results['shelf'] = await benchmarks.ShelfAppBenchmark().report();
   // results['vania (no_cli)'] = await benchmarks.VaniaAppBenchmark().report();
-  results['pharaoh'] = await benchmarks.PharaohAppBenchmark().report();
+  // results['pharaoh'] = await benchmarks.PharaohAppBenchmark().report();
   // results['angel3'] = await benchmarks.Angel3AppBenchmark().report();
-  results['dart_frog (no_cli)'] =
-      await benchmarks.DartFrogAppBenchmark().report();
+  // results['dart_frog (no_cli)'] =
+  //     await benchmarks.DartFrogAppBenchmark().report();
   results['dart_http'] = await benchmarks.DartHttpAppBenchmark().report();
   await saveToFile();
 }
 
 Future<void> saveToFile() async {
   final file = File('all_results.md');
+  final jsonFile = File('all_results.json');
   final lines = [
-    ' |                | Req/sec | Trans/sec | Req/sec DIFF | Avg Latency |',
-    ' |----------------|---------|-----------|-------------|-----------|'
+    ' | Server | Req/sec | Trans/sec | Req/sec DIFF | Avg Latency |',
+    ' |:-|:-|:-|:-|:-|'
   ];
+  final objs = [];
   final sorted = Map<String, Result?>.fromEntries(results.entries.toList()
     ..sort((a, b) => a.value!.rps.compareTo(b.value!.rps)));
   for (var entry in sorted.entries) {
@@ -37,6 +40,13 @@ Future<void> saveToFile() async {
     ;
     lines.add(
         ' | ${entry.key} | ${entry.value!.rps} | ${entry.value!.transferRate} | ${reqSecDiff.sign == -1.0 ? '' : '+'}${reqSecDiff.toStringAsFixed(2)}% | ${entry.value?.avgLatency} |');
+    objs.add({
+      'server': entry.key,
+      'rps': entry.value!.rps,
+      'transferRate': entry.value!.transferRate,
+      'reqSecDiff': reqSecDiff,
+      'avgLatency': entry.value?.avgLatency,
+    });
     final entryLines = [
       '## ${entry.key}',
       '### Requests per second',
@@ -55,6 +65,8 @@ Future<void> saveToFile() async {
       'Rps Perc: ${entry.value?.rpsPerc}',
     ];
     await entryFile.writeAsString(entryLines.join('\n'));
+    
   }
   await file.writeAsString(lines.join('\n'));
+  await jsonFile.writeAsString(jsonEncode({'results': objs}));
 }