+++ packages/serinus/lib/src/containers/module_container.dart
@@ -73,8 +73,6 @@ final class ModulesContainer {
           moduleInjectables!.concatTo(_moduleInjectables[token]);
     } else {
       final newInjectables = ModuleInjectables(
-        guards: {...module.guards},
-        pipes: {...module.pipes},
         middlewares: {...module.middlewares},
       );
       _moduleInjectables[token] =
@@ -140,8 +138,6 @@ final class ModulesContainer {
     final token = moduleToken(module);
     final currentModuleInjectables =
         _moduleInjectables[token] ??= ModuleInjectables(
-      guards: {...module.guards},
-      pipes: {...module.pipes},
       middlewares: {...module.middlewares},
     );
     for (var subModule in eagerSubModules) {
@@ -331,22 +327,14 @@ final class ModulesContainer {
 
 /// The [ModuleInjectables] class is used to create the module injectables.
 class ModuleInjectables {
-  /// The [guards] property contains the guards of the module
-  final Set<Guard> guards;
-
-  /// The [pipes] property contains the pipes of the module
-  final Set<Pipe> pipes;
+  /// The [providers] property contains the providers of the module
+  final Set<Provider> providers;
 
   /// The [middlewares] property contains the middlewares of the module
   final Set<Middleware> middlewares;
 
-  /// The [providers] property contains the providers of the module
-  final Set<Provider> providers;
-
   /// The constructor of the [ModuleInjectables] class
   ModuleInjectables({
-    required this.guards,
-    required this.pipes,
     required this.middlewares,
     this.providers = const {},
   });
@@ -354,8 +342,6 @@ class ModuleInjectables {
   /// Concatenates the module injectables with another module injectables
   ModuleInjectables concatTo(ModuleInjectables? moduleInjectables) {
     return ModuleInjectables(
-      guards: guards..addAllIfAbsent(moduleInjectables?.guards ?? {}),
-      pipes: pipes..addAllIfAbsent(moduleInjectables?.pipes ?? {}),
       middlewares: middlewares
         ..addAllIfAbsent(moduleInjectables?.middlewares ?? {}),
       providers: providers..addAllIfAbsent(moduleInjectables?.providers ?? {}),
@@ -364,14 +350,10 @@ class ModuleInjectables {
 
   /// Copies the module injectables with the new values
   ModuleInjectables copyWith({
-    Set<Guard>? guards,
-    Set<Pipe>? pipes,
     Set<Middleware>? middlewares,
     Set<Provider>? providers,
   }) {
     return ModuleInjectables(
-      guards: guards ?? this.guards,
-      pipes: pipes ?? this.pipes,
       middlewares: middlewares ?? this.middlewares,
       providers: providers ?? this.providers,
     );